# internal/domain/auth/handler.go
package auth

import (
	"log"
	"net/http"

	errors "github.com/RuanHOliveira/estatehub_api/internal/core/error"
	"github.com/RuanHOliveira/estatehub_api/internal/core/json"
)

type AuthHandler struct {
	u AuthUsecase
}

func NewAuthHandler(u AuthUsecase) *AuthHandler {
	return &AuthHandler{u: u}
}

// Register godoc
// @Summary      Registrar novo usuário
// @Description  Cria uma nova conta e retorna o token JWT de acesso
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        request body RegisterRequest true "Dados de registro"
// @Success      201 {object} RegisterResponse
// @Failure      400 {object} json.ErrorResponse
// @Failure      409 {object} json.ErrorResponse
// @Failure      500 {object} json.ErrorResponse
// @Router       /auth/register [post]
func (h *AuthHandler) Register(w http.ResponseWriter, r *http.Request) {
	var req RegisterRequest
	if err := json.Read(r, &req); err != nil {
		log.Println(err)
		json.WriteError(w, http.StatusBadRequest, errors.ErrInvalidRequest)
		return
	}

	output, err := h.u.Register(r.Context(), &RegisterInput{Email: req.Email, Name: req.Name, Password: req.Password})
	if err != nil {
		log.Println(err)

		if err == errors.ErrEmailAlreadyUsed {
			json.WriteError(w, http.StatusConflict, err)
			return
		}

		json.WriteError(w, http.StatusInternalServerError, errors.ErrUnknown)
		return
	}

	json.Write(w, http.StatusCreated, RegisterResponse{User: output.User, AccessToken: output.AccessToken})
}

// Login godoc
// @Summary      Autenticar usuário
// @Description  Valida credenciais e retorna o token JWT de acesso
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        request body LoginRequest true "Credenciais de login"
// @Success      200 {object} LoginResponse
// @Failure      400 {object} json.ErrorResponse
// @Failure      401 {object} json.ErrorResponse
// @Failure      404 {object} json.ErrorResponse
// @Failure      500 {object} json.ErrorResponse
// @Router       /auth/login [post]
func (h *AuthHandler) Login(w http.ResponseWriter, r *http.Request) {
	var req LoginRequest
	if err := json.Read(r, &req); err != nil {
		log.Println(err)
		json.WriteError(w, http.StatusBadRequest, errors.ErrInvalidRequest)
		return
	}

	output, err := h.u.Login(r.Context(), &LoginInput{Email: req.Email, Password: req.Password})
	if err != nil {
		if err == errors.ErrUserNotFound {
			json.WriteError(w, http.StatusNotFound, err)
			return
		}

		if err == errors.ErrInvalidCredentials {
			json.WriteError(w, http.StatusUnauthorized, err)
			return
		}

		json.WriteError(w, http.StatusInternalServerError, errors.ErrUnknown)
		return
	}

	json.Write(w, http.StatusOK, LoginResponse{User: output.User, AccessToken: output.AccessToken})
}